+++ .github/workflows/check-lockfile.yaml
@@ -1,0 +1,22 @@
+name: Check if the lockfile is up-to-date
+
+on:
+  pull_request:
+    branches: [main]
+  push:
+    branches: [main]
+
+jobs:
+  lockfile-up-to-date:
+    name: Check if the lockfile is up-to-date
+    runs-on: ubuntu-latest
+
+    steps:
+    - name: Checkout repository
+      uses: actions/checkout@v4
+
+    - uses: prefix-dev/setup-pixi@v0.8.1
+      with:
+        pixi-version: latest
+        cache: true
+        locked: true
