+++ .github/workflows/label_sync.yaml
@@ -1,0 +1,17 @@
+name: Sync Closing Labels
+on:
+- pull_request_target
+
+jobs:
+  sync:
+    permissions:
+      contents: read
+      pull-requests: write
+    runs-on: ubuntu-latest
+    steps:
+    - name: Checkout repository
+      uses: actions/checkout@v4
+    - name: Sync labels with closing issues
+      uses: wd60622/closing-labels@v0.0.4
+      env:
+        GH_TOKEN: ${{ github.token }}
