--- requirements.txt
@@ -1,17 +1,0 @@
-jupyterlab
-numpy
-pandas
-matplotlib
-scikit-learn
-langchain-core
-langchain-community>=0.0.13
-langchain-openai>=0.0.5
-langchain-anthropic
-langgraph
-smolagents
-e2b-code-interpreter==1.2.0b1
-tenacity
-discord.py>=2.3.2
-python-dotenv>=1.0.0
-colorama>=0.4.6
-slack-sdk>=3.26.1
