+++ what_the_chat/llm/__init__.py
@@ -1,0 +1,6 @@
+"""LLM services for summarization and interactive chat."""
+
+from .chat import ChatService
+from .summarization import SummarizationService
+
+__all__ = ["ChatService", "SummarizationService"]
