+++ what_the_chat/models/__init__.py
@@ -1,0 +1,5 @@
+"""Data models for chat messages and history."""
+
+from .message import ChatMessage, ChatHistory
+
+__all__ = ["ChatMessage", "ChatHistory"]
