+++ what_the_chat/platforms/__init__.py
@@ -1,0 +1,6 @@
+"""Platform integrations for different chat services."""
+
+from .discord import DiscordPlatform
+from .slack import SlackPlatform
+
+__all__ = ["DiscordPlatform", "SlackPlatform"]
