+++ what_the_chat/utils/__init__.py
@@ -1,0 +1,5 @@
+"""Utility functions for text processing and formatting."""
+
+from .formatting import standardize_user_references, replace_user_ids_with_names
+
+__all__ = ["standardize_user_references", "replace_user_ids_with_names"]
